pyproject.toml
[project]
name = "airflow-toolkit"
version = "0.13.3"
description = ""
authors = [{ name = "Biel Llobera", email = "biel_llobera@dkl.digital" }]
requires-python = ">=3.9,<3.12"
readme = "README.md"
dependencies = [
    "airflow-provider-duckdb>=0.2.0,<0.3",
    "apache-airflow>=2.8.0,<3",
    "apache-airflow-providers-amazon>=9.0.0,<10",
    "apache-airflow-providers-google>=10.23.0,<11",
    "apache-airflow-providers-microsoft-azure>=8.0.0,<9",
    "apache-airflow-providers-sftp>=4.8.1,<5",
    "apache-airflow-providers-slack>=8.5.1,<9",
    "boto3>=1.34.8,<2",
    "boto3-stubs>=1.34.8,<2",
    "databricks-sdk>=0.21.0,<0.22",
    "databricks-sql-connector>=2",
    "deltalake>=0.17.4,<0.18",
    "flask-session==0.5.0",
    "jmespath>=1.0.1,<2",
    "pandas>=2.1.1,<3",
    "pandas-stubs>=2.1.4.231218,<3",
    "pendulum>=2.1.2,<3",
    "sqlalchemy==1.4.54",
    "toml>=0.10.2,<0.11",
    "types-jmespath>=1.0.2.7,<2",
    "types-toml>=0.10.8.7,<0.11",
    "paramiko<4.0.0",
]

[project.entry-points.apache_airflow_provider]
provider_info = "airflow_toolkit.providers.package:get_provider_info"

[dependency-groups]
dev = [
    "black>=23.10.0,<24",
    "ruff>=0.1.0,<0.2",
    "isort>=5.12.0,<6",
    "mypy>=1.8.0,<2",
    "ipdb>=0.13.13,<0.14",
    "toml>=0.10.2",
]
test = [
    "pytest>=7.4.2,<8",
    "pytest-cov>=4.1.0,<5",
    "werkzeug<3.0.0",
    "connexion>=2.14.2,<3",
    "boto3>=1.28.67,<2",
    "pytest-mock>=3.12.0,<4",
    "requests-mock>=1.12.1,<2",
]

[tool.uv]
default-groups = [
    "dev",
    "test",
]

[tool.hatch.build.targets.sdist]
include = ["src/airflow_toolkit"]

[tool.hatch.build.targets.wheel]
include = ["src/airflow_toolkit"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.isort]
profile = "black"

[tool.black]
skip-string-normalization = true

[tool.mypy]
allow_redefinition = true
